// frontend/src/types/analytics.ts
export type BestSellerItem = {
  title: string
  sales_count: number
  total_sales_jpy: number
  total_profit_jpy: number
  avg_profit_margin: number
  avg_days_to_sell: number
}

export type CategoryAnalyticsItem = {
  category: string
  total_sales_jpy: number
  total_profit_jpy: number
  avg_profit_margin: number
  avg_days_to_sell: number
}